Move GDE related activities/fragments to gde package
